% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Rscript_filename}
\alias{get_Rscript_filename}
\title{Returns file name being interpreted by Rscript}
\usage{
get_Rscript_filename()
}
\value{
A string with the filename of the calling script.
If not found (i.e. you are in a interactive session) returns \code{NA_character_}.
}
\description{
\code{get_Rscript_filename()} returns the file name that \code{Rscript} is interpreting.
}
