% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{getoperand}
\alias{getoperand}
\title{Extract positional arguments from a getopt result}
\usage{
getoperand(x)
}
\arguments{
\item{x}{An object of class \code{"getopt"} as returned by \code{\link[=getopt]{getopt()}}.}
}
\value{
A character vector of positional arguments.
}
\description{
Extracts the positional arguments stored in the \code{"operands"} attribute of the
object returned by \code{\link[=getopt]{getopt()}}.
}
\examples{
spec <- matrix(c("verbose", "v", 0, "logical"), ncol = 4, byrow = TRUE)
opt <- getopt(spec, c("--verbose", "--", "file1.txt", "file2.txt"))
getoperand(opt)
}
