% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{getusage}
\alias{getusage}
\title{Generate a usage string}
\usage{
getusage(spec, command = get_Rscript_filename())
}
\arguments{
\item{spec}{The getopt specification, or spec of what options are considered
valid.  The specification must be either a 4-5 column matrix, or a
character vector coercible into a 4 column matrix using
\code{matrix(x,ncol=4,byrow=TRUE)} command.  The matrix/vector
contains:

Column 1: the \emph{long flag} name.  A multi-character string.

Column 2: \emph{short flag} alias of Column 1.  A single-character
string.

Column 3: \emph{Argument} mask of the \emph{flag}.  An integer.
Possible values: 0=no argument, 1=required argument, 2=optional argument.

Column 4: Data type to which the \emph{flag}'s argument shall be cast using
\code{\link[=storage.mode]{storage.mode()}}.  A multi-character string.  This only considered
for same-row Column 3 values of 1,2.  Possible values: logical,
integer, double, complex, character.
If numeric is encountered then it will be converted to double.

Column 5 (optional): A brief description of the purpose of the option.

The terms \emph{option}, \emph{flag}, \emph{long flag}, \emph{short flag},
and \emph{argument} have very specific meanings in the context of this
document.  Read the \dQuote{Description} section for definitions.}

\item{command}{The string to use in the usage message as the name of the
script.  See argument \emph{usage}.}
}
\value{
A character string with the usage message.
}
\description{
Generate a usage string from a getopt \code{spec} matrix.
}
\examples{
spec <- matrix(c(
  'verbose', 'v', 2, "integer",
  'help'   , 'h', 0, "logical",
  'count'  , 'c', 1, "integer",
  'mean'   , 'm', 1, "double",
  'sd'     , 's', 1, "double"
), byrow = TRUE, ncol = 4)
cat(getusage(spec, command = "myscript"))
}
