% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_list}
\alias{sort_list}
\title{Recursively sorts a list}
\usage{
sort_list(unsorted_list)
}
\arguments{
\item{unsorted_list}{A list.}
}
\value{
A sorted list.
}
\description{
\code{sort_list} returns a sorted list
}
